Move parse_interpolation() to Transform
