+++ torchio/transforms/augmentation/random_transform.py
@@ -4,14 +4,13 @@ This is the docstring of random transform module
 
 import numbers
 from typing import Optional, Tuple, Union
-import warnings
 
 import torch
 import numpy as np
 
 from ...data.subject import Subject
 from ... import TypeNumber, TypeRangeFloat
-from .. import Transform, Interpolation
+from .. import Transform
 
 
 class RandomTransform(Transform):
@@ -84,23 +83,6 @@ class RandomTransform(Transform):
             translation: TypeRangeFloat,
             ) -> Tuple[float, float]:
         return self.parse_range(translation, 'translation')
-
-    @staticmethod
-    def parse_interpolation(interpolation: str) -> Interpolation:
-        if isinstance(interpolation, Interpolation):
-            message = 'Interpolation of type torchio.Interpolation is deprecated, please use a String instead.'
-            warnings.warn(message, FutureWarning)
-        elif isinstance(interpolation, str):
-            supported_values = [key.name.lower() for key in Interpolation]
-            if interpolation in supported_values:
-                interpolation = getattr(Interpolation, interpolation.upper())
-            else:
-                message = f'Interpolation {interpolation} is not among torchio supported values: {supported_values}'
-                raise AttributeError(message)
-        else:
-            message = 'image_interpolation must be a String'
-            raise TypeError(message)
-        return interpolation
 
     def check_seed(self) -> None:
         if self._seed is not None:
